requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
click==6.7
Flask==0.10.1
flask-mongoengine==0.9.2
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mongoengine==0.12.0
olefile==0.44
packaging==16.8
Pillow==4.1.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
scikit-learn==0.18.1
six==1.10.0
Werkzeug==0.11.9
WTForms==2.1
mongoengine==0.12.0